Add loot distribution to adventure outcomes
Added functionality to distribute loot items to players during adventures. This is implemented through a `maybe_loot` function that uses a random choice to determine whether a loot item should be distributed or not. If an item is to be distributed and the player already owns the item, the quantity is increased. If the player does not have the item, a new item is created and added to the player's items. The changes also involved moving items data from `lists.py` to a new file `items.py`.