Implement new Help command and HelpView interface
Added a new Help command and a HelpView interface to the bot's core commands. This change offers an interactive and more user-friendly way of accessing the bot's command help. Instead of a large embed containing all commands, users will select a category from a dropdown menu and get an embed with the related commands.