Add method to fetch user or display 'Unknown'
A new method 'get_name_or_unknown' has been added to the profile.py. This method attempts to fetch a user based on the provided ID and returns "Unknown" if it either fails or the ID is absent in the available data. This function is now used to get the 'owner_name' and 'likes' fields.