Restrict bot commands to guild only in core.py
Bot commands in the core.py file have been restricted to be executed in guild context only. This was achieved by adding the `@commands.guild_only()` decorator to each command. With this change, the bot commands won't function in direct messages but will only be available in server/guild channels, enhancing the usability in a guild context.