Add new file for handling stats view
The newly added file, stats_view.py, creates a 'Stats' class for Discord user interaction related to pulling user statistics. The "StatsView" class is used for displaying options to the users for different types of stats they want to view. This addition helps in bringing a better and more interactive experience for the users while viewing their stats.