Revise bail logic to support third-party payments.
This update allows a user to bail out another user by checking the author's balance and deducting from it. If the author does not have sufficient funds, the bail process is halted with a notification. The changes ensure that bailing can be executed by a third party, enhancing interactivity within the platform.